Modificaciones para el jpa, MascotasController no funcional

+++ demo/src/main/java/com/example/demo/servicio/MascotaService.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Service;
 import com.example.demo.entidades.Mascota;
 
 public interface MascotaService {
-    public Mascota searchById(int id);
+    public Mascota searchById(Long id);
 
     public Collection<Mascota> searchAll();
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 
     public void update(Mascota mascota);
